Authorization engines evaluate privileges for APIs
This commit moves the evaluation of privileges from a few transport
actions into the authorization engine. The APIs are used by other
applications for making decisions and if a different authorization
engine is used that is not role based, we should still allow these APIs
to work. By moving this evaluation out of the transport action, the
transport actions no longer have a dependency on roles.
