+++ x-pack/plugin/security/src/main/java/org/elasticsearch/xpack/security/action/user/TransportHasPrivilegesAction.java
@@ -5,15 +5,10 @@
  */
 package org.elasticsearch.xpack.security.action.user;
 
-import org.apache.logging.log4j.message.ParameterizedMessage;
-import org.apache.lucene.util.automaton.Automaton;
-import org.apache.lucene.util.automaton.Operations;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.support.ActionFilters;
 import org.elasticsearch.action.support.HandledTransportAction;
-import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.inject.Inject;
-import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.tasks.Task;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportService;
@@ -22,27 +17,13 @@ import org.elasticsearch.xpack.core.security.action.user.HasPrivilegesRequest;
 import org.elasticsearch.xpack.core.security.action.user.HasPrivilegesResponse;
 import org.elasticsearch.xpack.core.security.authc.Authentication;
 import org.elasticsearch.xpack.core.security.authz.RoleDescriptor;
-import org.elasticsearch.xpack.core.security.authz.permission.FieldPermissionsCache;
-import org.elasticsearch.xpack.core.security.authz.permission.IndicesPermission;
-import org.elasticsearch.xpack.core.security.authz.permission.Role;
-import org.elasticsearch.xpack.core.security.authz.privilege.ApplicationPrivilege;
 import org.elasticsearch.xpack.core.security.authz.privilege.ApplicationPrivilegeDescriptor;
-import org.elasticsearch.xpack.core.security.authz.privilege.ClusterPrivilege;
-import org.elasticsearch.xpack.core.security.authz.privilege.IndexPrivilege;
-import org.elasticsearch.xpack.core.security.authz.privilege.Privilege;
-import org.elasticsearch.xpack.core.security.support.Automatons;
 import org.elasticsearch.xpack.core.security.user.User;
-import org.elasticsearch.xpack.security.authz.store.CompositeRolesStore;
+import org.elasticsearch.xpack.security.authz.AuthorizationService;
 import org.elasticsearch.xpack.security.authz.store.NativePrivilegeStore;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -53,16 +34,16 @@ import java.util.stream.Collectors;
 public class TransportHasPrivilegesAction extends HandledTransportAction<HasPrivilegesRequest, HasPrivilegesResponse> {
 
     private final ThreadPool threadPool;
-    private final CompositeRolesStore rolesStore;
+    private final AuthorizationService authorizationService;
     private final NativePrivilegeStore privilegeStore;
 
     @Inject
     public TransportHasPrivilegesAction(ThreadPool threadPool, TransportService transportService,
-                                        ActionFilters actionFilters, CompositeRolesStore rolesStore,
+                                        ActionFilters actionFilters, AuthorizationService authorizationService,
                                         NativePrivilegeStore privilegeStore) {
         super(HasPrivilegesAction.NAME, transportService, actionFilters, HasPrivilegesRequest::new);
         this.threadPool = threadPool;
-        this.rolesStore = rolesStore;
+        this.authorizationService = authorizationService;
         this.privilegeStore = privilegeStore;
     }
 
@@ -70,17 +51,15 @@ public class TransportHasPrivilegesAction extends HandledTransportAction<HasPriv
     protected void doExecute(Task task, HasPrivilegesRequest request, ActionListener<HasPrivilegesResponse> listener) {
         final String username = request.username();
 
-        final User user = Authentication.getAuthentication(threadPool.getThreadContext()).getUser();
+        final Authentication authentication = Authentication.getAuthentication(threadPool.getThreadContext());
+        final User user = authentication.getUser();
         if (user.principal().equals(username) == false) {
             listener.onFailure(new IllegalArgumentException("users may only check the privileges of their own account"));
             return;
         }
 
-        // FIXME
-        rolesStore.getRoles(user, new FieldPermissionsCache(Settings.EMPTY), ActionListener.wrap(
-            role -> resolveApplicationPrivileges(request, ActionListener.wrap(
-                applicationPrivilegeLookup -> checkPrivileges(request, role, applicationPrivilegeLookup, listener),
-                listener::onFailure)),
+        resolveApplicationPrivileges(request, ActionListener.wrap(applicationPrivilegeDescriptors ->
+                authorizationService.checkPrivileges(authentication, request, applicationPrivilegeDescriptors, listener),
             listener::onFailure));
     }
 
@@ -90,125 +69,9 @@ public class TransportHasPrivilegesAction extends HandledTransportAction<HasPriv
         privilegeStore.getPrivileges(applications, null, listener);
     }
 
-    private Set<String> getApplicationNames(HasPrivilegesRequest request) {
+    public static Set<String> getApplicationNames(HasPrivilegesRequest request) {
         return Arrays.stream(request.applicationPrivileges())
             .map(RoleDescriptor.ApplicationResourcePrivileges::getApplication)
             .collect(Collectors.toSet());
     }
-
-    private void checkPrivileges(HasPrivilegesRequest request, Role userRole,
-                                 Collection<ApplicationPrivilegeDescriptor> applicationPrivileges,
-                                 ActionListener<HasPrivilegesResponse> listener) {
-        logger.trace(() -> new ParameterizedMessage("Check whether role [{}] has privileges cluster=[{}] index=[{}] application=[{}]",
-            Strings.arrayToCommaDelimitedString(userRole.names()),
-            Strings.arrayToCommaDelimitedString(request.clusterPrivileges()),
-            Strings.arrayToCommaDelimitedString(request.indexPrivileges()),
-            Strings.arrayToCommaDelimitedString(request.applicationPrivileges())
-        ));
-
-        Map<String, Boolean> cluster = new HashMap<>();
-        for (String checkAction : request.clusterPrivileges()) {
-            final ClusterPrivilege checkPrivilege = ClusterPrivilege.get(Collections.singleton(checkAction));
-            final ClusterPrivilege rolePrivilege = userRole.cluster().privilege();
-            cluster.put(checkAction, testPrivilege(checkPrivilege, rolePrivilege.getAutomaton()));
-        }
-        boolean allMatch = cluster.values().stream().allMatch(Boolean::booleanValue);
-
-        final Map<IndicesPermission.Group, Automaton> predicateCache = new HashMap<>();
-
-        final Map<String, HasPrivilegesResponse.ResourcePrivileges> indices = new LinkedHashMap<>();
-        for (RoleDescriptor.IndicesPrivileges check : request.indexPrivileges()) {
-            for (String index : check.getIndices()) {
-                final Map<String, Boolean> privileges = new HashMap<>();
-                final HasPrivilegesResponse.ResourcePrivileges existing = indices.get(index);
-                if (existing != null) {
-                    privileges.putAll(existing.getPrivileges());
-                }
-                for (String privilege : check.getPrivileges()) {
-                    if (testIndexMatch(index, check.allowRestrictedIndices(), privilege, userRole, predicateCache)) {
-                        logger.debug(() -> new ParameterizedMessage("Role [{}] has [{}] on index [{}]",
-                            Strings.arrayToCommaDelimitedString(userRole.names()), privilege, index));
-                        privileges.put(privilege, true);
-                    } else {
-                        logger.debug(() -> new ParameterizedMessage("Role [{}] does not have [{}] on index [{}]",
-                            Strings.arrayToCommaDelimitedString(userRole.names()), privilege, index));
-                        privileges.put(privilege, false);
-                        allMatch = false;
-                    }
-                }
-                indices.put(index, new HasPrivilegesResponse.ResourcePrivileges(index, privileges));
-            }
-        }
-
-        final Map<String, Collection<HasPrivilegesResponse.ResourcePrivileges>> privilegesByApplication = new HashMap<>();
-        for (String applicationName : getApplicationNames(request)) {
-            logger.debug("Checking privileges for application {}", applicationName);
-            final Map<String, HasPrivilegesResponse.ResourcePrivileges> appPrivilegesByResource = new LinkedHashMap<>();
-            for (RoleDescriptor.ApplicationResourcePrivileges p : request.applicationPrivileges()) {
-                if (applicationName.equals(p.getApplication())) {
-                    for (String resource : p.getResources()) {
-                        final Map<String, Boolean> privileges = new HashMap<>();
-                        final HasPrivilegesResponse.ResourcePrivileges existing = appPrivilegesByResource.get(resource);
-                        if (existing != null) {
-                            privileges.putAll(existing.getPrivileges());
-                        }
-                        for (String privilege : p.getPrivileges()) {
-                            if (testResourceMatch(applicationName, resource, privilege, userRole, applicationPrivileges)) {
-                                logger.debug(() -> new ParameterizedMessage("Role [{}] has [{} {}] on resource [{}]",
-                                    Strings.arrayToCommaDelimitedString(userRole.names()), applicationName, privilege, resource));
-                                privileges.put(privilege, true);
-                            } else {
-                                logger.debug(() -> new ParameterizedMessage("Role [{}] does not have [{} {}] on resource [{}]",
-                                    Strings.arrayToCommaDelimitedString(userRole.names()), applicationName, privilege, resource));
-                                privileges.put(privilege, false);
-                                allMatch = false;
-                            }
-                        }
-                        appPrivilegesByResource.put(resource, new HasPrivilegesResponse.ResourcePrivileges(resource, privileges));
-                    }
-                }
-            }
-            privilegesByApplication.put(applicationName, appPrivilegesByResource.values());
-        }
-
-        listener.onResponse(new HasPrivilegesResponse(request.username(), allMatch, cluster, indices.values(), privilegesByApplication));
-    }
-
-    private boolean testIndexMatch(String checkIndexPattern, boolean allowRestrictedIndices, String checkPrivilegeName, Role userRole,
-            Map<IndicesPermission.Group, Automaton> predicateCache) {
-        final IndexPrivilege checkPrivilege = IndexPrivilege.get(Collections.singleton(checkPrivilegeName));
-
-        final Automaton checkIndexAutomaton = IndicesPermission.Group.buildIndexMatcherAutomaton(allowRestrictedIndices, checkIndexPattern);
-
-        List<Automaton> privilegeAutomatons = new ArrayList<>();
-        for (IndicesPermission.Group group : userRole.indices().groups()) {
-            final Automaton groupIndexAutomaton = predicateCache.computeIfAbsent(group,
-                    g -> IndicesPermission.Group.buildIndexMatcherAutomaton(g.allowRestrictedIndices(), g.indices()));
-            if (Operations.subsetOf(checkIndexAutomaton, groupIndexAutomaton)) {
-                final IndexPrivilege rolePrivilege = group.privilege();
-                if (rolePrivilege.name().contains(checkPrivilegeName)) {
-                    return true;
-                }
-                privilegeAutomatons.add(rolePrivilege.getAutomaton());
-            }
-        }
-        return testPrivilege(checkPrivilege, Automatons.unionAndMinimize(privilegeAutomatons));
-    }
-
-    private static boolean testPrivilege(Privilege checkPrivilege, Automaton roleAutomaton) {
-        return Operations.subsetOf(checkPrivilege.getAutomaton(), roleAutomaton);
-    }
-
-    private boolean testResourceMatch(String application, String checkResource, String checkPrivilegeName, Role userRole,
-                                      Collection<ApplicationPrivilegeDescriptor> privileges) {
-        final Set<String> nameSet = Collections.singleton(checkPrivilegeName);
-        final ApplicationPrivilege checkPrivilege = ApplicationPrivilege.get(application, nameSet, privileges);
-        assert checkPrivilege.getApplication().equals(application)
-            : "Privilege " + checkPrivilege + " should have application " + application;
-        assert checkPrivilege.name().equals(nameSet)
-            : "Privilege " + checkPrivilege + " should have name " + nameSet;
-
-        return userRole.application().grants(checkPrivilege, checkResource);
-    }
-
 }
